refactor(test-gen): custom format for enums
