+++ Cargo.toml
@@ -3,7 +3,8 @@ resolver = "2"
 members = [
     "rasn-compiler",
     "rasn-compiler-derive",
-    "rasn-compiler-tests"
+    "rasn-compiler-tests",
+    "rasn-compiler/internal-macros"
 ]
 
 [workspace.package]
