+++ rasn-compiler/Cargo.toml
@@ -32,6 +32,7 @@ cli = ["clap", "colored", "walkdir"]
 chrono = "0.4.31"
 clap = { version = "4.4.4", optional = true, features = ["derive"] }
 colored = { version = "2", optional = true }
+internal_macros = { path = "internal-macros" }
 nom = "7.1.3"
 proc-macro2 = "1.0"
 quote = "1.0"
