+++ rasn-compiler/internal-macros/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "internal_macros"
+publish = false
+description = "Internal derive macros for rasn."
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { version = "1.0.39", features = ["extra-traits"] }
+quote = "1.0.7"
+proc-macro2 = "1.0.19"
